Fix ability legalization error on bad gen4/5 origin
hacked initial kakuna yields -1 for PIDAbility, but pops out as it's
returned via AbilityNumber's override, keep sanity checking...

+++ PKHeX.Core.AutoMod/BruteForce/BruteForce.cs
@@ -300,7 +300,7 @@ namespace PKHeX.Core.AutoMod
             for (int i = 0; i < 3; i++)
             {
                 pk.AbilityNumber = abilityNumList[i];
-                pk.RefreshAbility(pk.AbilityNumber < 6 ? pk.AbilityNumber >> 1 : 0);
+                pk.RefreshAbility((uint)pk.AbilityNumber < 6 ? pk.AbilityNumber >> 1 : 0);
                 if (pk.Ability == abilityID)
                 {
                     var updatedReport = GetReport(pk);
@@ -312,7 +312,7 @@ namespace PKHeX.Core.AutoMod
                 }
             }
             pk.AbilityNumber = finalabilitynum;
-            pk.RefreshAbility(pk.AbilityNumber < 6 ? pk.AbilityNumber >> 1 : 0);
+            pk.RefreshAbility((uint)pk.AbilityNumber < 6 ? pk.AbilityNumber >> 1 : 0);
         }
 
         private static bool CommonErrorHandling2(PKM pk)
